feat: merge checklist task into check item
