--- backend/plugins/trello/e2e/snapshot_tables/_tool_trello_check_items.csv
@@ -1,57 +1,0 @@
-id,name,state,id_checklist,pos
-6402f644d23aa9af56b2928a,[Example task],incomplete,6402f643d23aa9af56b29019,33751
-6402f644d23aa9af56b2928b,[Another example task],incomplete,6402f643d23aa9af56b29019,50552
-6402f644d23aa9af56b29290,[Example task],complete,6402f643d23aa9af56b2901a,17309
-6402f644d23aa9af56b29291,[Another example task],complete,6402f643d23aa9af56b2901a,34469
-6402f644d23aa9af56b29296,Filter by date tweeted,incomplete,6402f643d23aa9af56b2901b,17265
-6402f644d23aa9af56b29297,Create a form to generate tweet report,incomplete,6402f643d23aa9af56b2901b,8632.5
-6402f644d23aa9af56b29298,Implement report functionality,incomplete,6402f643d23aa9af56b2901b,12948.75
-6402f644d23aa9af56b29299,Download report as CSV,incomplete,6402f643d23aa9af56b2901b,34191
-6402f644d23aa9af56b2929a,Download report as PDF,incomplete,6402f643d23aa9af56b2901b,50702
-6402f644d23aa9af56b292a8,Create form to register a new user,incomplete,6402f643d23aa9af56b2901f,17126
-6402f644d23aa9af56b292a9,Implement functionality to register a new user,incomplete,6402f643d23aa9af56b2901f,33572
-6402f644d23aa9af56b292aa,Implement authentication endpoint for mobile app developer,incomplete,6402f643d23aa9af56b2901f,17304
-6402f644d23aa9af56b292ab,Implement endpoint to register new user,incomplete,6402f643d23aa9af56b2901f,17215
-6402f644d23aa9af56b292b2,Document endpoint on postman,incomplete,6402f643d23aa9af56b29020,33845
-6402f644d23aa9af56b292b6,Upload endpoint returns a 400 error code,incomplete,6402f643d23aa9af56b29021,17152
-6402f644d23aa9af56b292ba,Implement endpoint to upload file,complete,6402f643d23aa9af56b29022,16661
-6402f644d23aa9af56b292bb,Implement endpoint to validate file,complete,6402f643d23aa9af56b29022,34005
-6402f644d23aa9af56b292bc,Implement endpoint to tag files in folders,complete,6402f643d23aa9af56b29022,67329
-6402f644d23aa9af56b292bd,Implement endpoint to store file on cloudinary,complete,6402f643d23aa9af56b29022,132865
-6402f644d23aa9af56b292be,Create a form to send upload request,incomplete,6402f643d23aa9af56b29022,198401
-6402f644d23aa9af56b292c6,Implement functionality to send a new tweet,incomplete,6402f643d23aa9af56b29023,33572
-6402f644d23aa9af56b292ca,Document endpoint on postman,complete,6402f643d23aa9af56b29024,33845
-6402f644d23aa9af56b292cb,Implement endpoint to send new tweet,complete,6402f643d23aa9af56b29024,25574.5
-6402f644d23aa9af56b292cc,Create form to send a new tweet,incomplete,6402f643d23aa9af56b29024,99381
-6402f644d23aa9af56b292d2,Create like button,incomplete,6402f643d23aa9af56b29026,17126
-6402f644d23aa9af56b292d3,Implement functionality to like tweet,incomplete,6402f643d23aa9af56b29026,33572
-6402f644d23aa9af56b292d8,Document endpoint on postman,incomplete,6402f643d23aa9af56b29025,33845
-6402f644d23aa9af56b292d9,Implement endpoint to like tweet,complete,6402f643d23aa9af56b29025,25574.5
-6402f644d23aa9af56b292de,[Example task],incomplete,6402f643d23aa9af56b29027,33751
-6402f644d23aa9af56b292df,[Another example task],incomplete,6402f643d23aa9af56b29027,50552
-6402f644d23aa9af56b292e4,[Example task],incomplete,6402f643d23aa9af56b29028,17309
-6402f644d23aa9af56b292e5,[Another example task],incomplete,6402f643d23aa9af56b29028,34469
-6402f644d23aa9af56b292ea,[Example task],incomplete,6402f643d23aa9af56b2902a,33751
-6402f644d23aa9af56b292eb,[Another example task],incomplete,6402f643d23aa9af56b2902a,50552
-6402f644d23aa9af56b292f0,[Example task],incomplete,6402f643d23aa9af56b29029,17309
-6402f644d23aa9af56b292f1,[Another example task],incomplete,6402f643d23aa9af56b29029,34469
-6402f644d23aa9af56b292f6,[Example task],incomplete,6402f643d23aa9af56b2902b,33751
-6402f644d23aa9af56b292f7,[Another example task],incomplete,6402f643d23aa9af56b2902b,50552
-6402f644d23aa9af56b292fc,[Example task],incomplete,6402f643d23aa9af56b2902c,17309
-6402f644d23aa9af56b292fd,[Another example task],incomplete,6402f643d23aa9af56b2902c,34469
-6402f644d23aa9af56b29302,[Example task],incomplete,6402f643d23aa9af56b2902d,33751
-6402f644d23aa9af56b29303,[Another example task],incomplete,6402f643d23aa9af56b2902d,50552
-6402f644d23aa9af56b29308,[Example task],incomplete,6402f643d23aa9af56b2902e,17309
-6402f644d23aa9af56b29309,[Another example task],incomplete,6402f643d23aa9af56b2902e,34469
-6402f644d23aa9af56b2930e,[Example task],incomplete,6402f643d23aa9af56b2902f,33751
-6402f644d23aa9af56b2930f,[Another example task],incomplete,6402f643d23aa9af56b2902f,50552
-6402f644d23aa9af56b29314,[Example task],incomplete,6402f643d23aa9af56b29030,17309
-6402f644d23aa9af56b29315,[Another example task],incomplete,6402f643d23aa9af56b29030,34469
-6402f644d23aa9af56b2931a,[Example task],incomplete,6402f643d23aa9af56b29032,33751
-6402f644d23aa9af56b2931b,[Another example task],incomplete,6402f643d23aa9af56b29032,50552
-6402f644d23aa9af56b29320,[Example task],incomplete,6402f643d23aa9af56b29031,17309
-6402f644d23aa9af56b29321,[Another example task],incomplete,6402f643d23aa9af56b29031,34469
-6402f644d23aa9af56b29326,[Example task],incomplete,6402f643d23aa9af56b29033,33751
-6402f644d23aa9af56b29327,[Another example task],incomplete,6402f643d23aa9af56b29033,50552
-6402f644d23aa9af56b2932c,[Example task],incomplete,6402f643d23aa9af56b29034,17309
-6402f644d23aa9af56b2932d,[Another example task],incomplete,6402f643d23aa9af56b29034,34469
