--- backend/plugins/trello/e2e/snapshot_tables/_tool_trello_checklists.csv
@@ -1,29 +1,0 @@
-id,name,id_board,id_card,pos
-6402f643d23aa9af56b29019,To-Do List,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b28ffd,73728
-6402f643d23aa9af56b2901a,Task Review,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b28ffd,8192
-6402f643d23aa9af56b2901b,To-do List,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b28ffe,16384
-6402f643d23aa9af56b2901c,Task Review,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b28ffe,8192
-6402f643d23aa9af56b2901d,Task Review,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b28fff,8192
-6402f643d23aa9af56b2901e,To-do List,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b28fff,16384
-6402f643d23aa9af56b2901f,To-do List,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29000,16384
-6402f643d23aa9af56b29020,Task Review,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29000,8192
-6402f643d23aa9af56b29021,To-do List,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29001,16384
-6402f643d23aa9af56b29022,Task Review,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29001,8192
-6402f643d23aa9af56b29023,To-do List,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29002,16384
-6402f643d23aa9af56b29024,Task Review,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29002,8192
-6402f643d23aa9af56b29025,Task Review,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29003,8192
-6402f643d23aa9af56b29026,To-do List,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29003,16384
-6402f643d23aa9af56b29027,To-Do List,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29004,73728
-6402f643d23aa9af56b29028,Task Review,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29004,8192
-6402f643d23aa9af56b29029,Task Review,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29005,8192
-6402f643d23aa9af56b2902a,To-Do List,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29005,73728
-6402f643d23aa9af56b2902b,To-Do List,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29006,73728
-6402f643d23aa9af56b2902c,Task Review,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29006,8192
-6402f643d23aa9af56b2902d,To-Do List,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29007,73728
-6402f643d23aa9af56b2902e,Task Review,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29007,8192
-6402f643d23aa9af56b2902f,To-Do List,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29008,73728
-6402f643d23aa9af56b29030,Task Review,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29008,8192
-6402f643d23aa9af56b29031,Task Review,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29009,8192
-6402f643d23aa9af56b29032,To-Do List,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b29009,73728
-6402f643d23aa9af56b29033,To-Do List,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b2900a,73728
-6402f643d23aa9af56b29034,Task Review,6402f643d23aa9af56b28f4b,6402f643d23aa9af56b2900a,8192
