+++ backend/plugins/trello/models/check_item.go
@@ -22,11 +22,14 @@ import (
 )
 
 type TrelloCheckItem struct {
-	ID          string `gorm:"primaryKey;type:varchar(255)"`
-	Name        string `gorm:"type:varchar(255)"`
-	State       string `gorm:"type:varchar(255)"`
-	IDChecklist string `gorm:"type:varchar(255)"`
-	Pos         float64
+	ID            string `gorm:"primaryKey;type:varchar(255)"`
+	Name          string `gorm:"type:varchar(255)"`
+	State         string `gorm:"type:varchar(255)"`
+	IDChecklist   string `gorm:"type:varchar(255)"`
+	ChecklistName string `gorm:"type:varchar(255)"`
+	IDBoard       string `gorm:"type:varchar(255)"`
+	IDCard        string `gorm:"type:varchar(255)"`
+	Pos           float64
 	common.NoPKModel
 }
